feat: Implementa a lógica para exibir o conteúdo de acordo com o perfil (admin ou agent)

+++ app/views/homepage.php
@@ -1,33 +1,38 @@
 <div class="container-fluid mt-5">
     <div class="row justify-content-center">
-        
+
         <div class="d-flex flex-row flex-wrap justify-content-center">
-            
-            <!-- os meus clientes -->
-            <a href="#" class="unlink m-2">
-                <div class="home-option p-5 text-center">
-                    <h3 class="mb-3"><i class="fa-solid fa-users"></i></h3>
-                    <h5>Os meus clientes</h5>
-                </div>
-            </a>
-            
-            <!-- adicionar clientes -->
-            <a href="#" class="unlink m-2">
-                <div class="home-option p-5 text-center">
-                    <h3 class="mb-3"><i class="fa-solid fa-user-plus"></i></h3>
-                    <h5>Adicionar clientes</h5>
-                </div>
-            </a>
 
-            <!-- carregar ficheiro de clientes -->
-            <a href="#" class="unlink m-2">
-                <div class="home-option p-5 text-center">
-                    <h3 class="mb-3"><i class="fa-solid fa-upload"></i></h3>
-                    <h5>Carregar ficheiro</h5>
-                </div>
-            </a>
+            <!-- Lógica para carregar o conteúdo dos agentes -->
+            <?php if ($user->profile == 'agent'): ?>
+                <!-- os meus clientes -->
+                <a href="#" class="unlink m-2">
+                    <div class="home-option p-5 text-center">
+                        <h3 class="mb-3"><i class="fa-solid fa-users"></i></h3>
+                        <h5>Os meus clientes</h5>
+                    </div>
+                </a>
 
+                <!-- adicionar clientes -->
+                <a href="#" class="unlink m-2">
+                    <div class="home-option p-5 text-center">
+                        <h3 class="mb-3"><i class="fa-solid fa-user-plus"></i></h3>
+                        <h5>Adicionar clientes</h5>
+                    </div>
+                </a>
+
+                <!-- carregar ficheiro de clientes -->
+                <a href="#" class="unlink m-2">
+                    <div class="home-option p-5 text-center">
+                        <h3 class="mb-3"><i class="fa-solid fa-upload"></i></h3>
+                        <h5>Carregar ficheiro</h5>
+                    </div>
+                </a>
+            <?php endif; ?> <!-- FIM da Lógica para carregar o conteúdo dos agentes -->
+
+            <!-- Lógica para carregar o conteúdo do ADMIN -->
             <!-- clientes -->
+            <?php if($user->profile == 'admin'): ?>
             <a href="#" class="unlink m-2">
                 <div class="home-option p-5 text-center">
                     <h3 class="mb-3"><i class="fa-solid fa-users"></i></h3>
@@ -50,6 +55,7 @@
                     <h5>Gestão de utilizadores</h5>
                 </div>
             </a>
+            <?php endif;?> <!-- FIM Lógica para carregar o conteúdo do ADMIN -->
 
         </div>
